Fixed Scrolling On Launch
For new users, the scrolling views will stay at the top, so the isntructional text shows. Once the user has run through the watch app twice the behavior will change. Now when the scroll interface loads, it will scroll to the bill amount last selected and then to the suggested tip on the tip page. Issues #1