Updated Rows when Currency Symbol Changes
Added notifications so the Table Rows change and update their currency symbol when it gets changed in settings. It doesn't actually change when it changes in settings. The watch prevents UI updates when a UI is not visible on the screen. Instead, changing a setting notifies the interface controller that has the tables rows in it, that the setting has changed. Then in willActivate() the interface controller tells the rows to update.

Resolves issue #11